remove redunant function (#6220)
`getPrompt` implement is same for all the adapters and we are not doing
any provider specific processing. Removed the redunant methods.

## Test plan
No functional change.
