+++ vscode/src/autoedits/adapters/fireworks.ts
@@ -1,50 +1,9 @@
-import type { AutoEditsTokenLimit } from '@sourcegraph/cody-shared'
-import type * as vscode from 'vscode'
-import type {
-    AutocompleteContextSnippet,
-    DocumentContext,
-} from '../../../../lib/shared/src/completions/types'
 import { autoeditsLogger } from '../logger'
-import type { AutoeditsModelAdapter, ChatPrompt, PromptResponseData } from '../prompt-provider'
+import type { AutoeditsModelAdapter } from '../prompt-provider'
 import { getModelResponse } from '../prompt-provider'
 import type { AutoeditModelOptions } from '../prompt-provider'
-import { type CodeToReplaceData, SYSTEM_PROMPT, getBaseUserPrompt } from '../prompt-utils'
 
 export class FireworksAdapter implements AutoeditsModelAdapter {
-    getPrompt(
-        docContext: DocumentContext,
-        document: vscode.TextDocument,
-        position: vscode.Position,
-        context: AutocompleteContextSnippet[],
-        tokenBudget: AutoEditsTokenLimit
-    ): PromptResponseData {
-        const { codeToReplace, prompt: userPrompt } = getBaseUserPrompt(
-            docContext,
-            document,
-            position,
-            context,
-            tokenBudget
-        )
-        const prompt: ChatPrompt = [
-            {
-                role: 'system',
-                content: SYSTEM_PROMPT,
-            },
-            {
-                role: 'user',
-                content: userPrompt,
-            },
-        ]
-        return {
-            codeToReplace,
-            promptResponse: prompt,
-        }
-    }
-
-    postProcessResponse(codeToReplace: CodeToReplaceData, response: string): string {
-        return response
-    }
-
     async getModelResponse(option: AutoeditModelOptions): Promise<string> {
         try {
             const response = await getModelResponse(
